+++ core/src/Streamly/Internal/FileSystem/Path.hs
@@ -12,19 +12,25 @@
 -- the filepath package API. It has some differences from the filepath
 -- package:
 --
--- 1. Empty paths are not allowed. Paths are validated before construction.
--- 2. The default Path type itself affords considerable safety regarding the
+-- 1. The append operations follows path construction semantics rather than
+-- path resolution and navigation based semantics used by the </> operation in
+-- filepath package. Better have run time failures instead of silent problems.
+-- 2. Empty paths are not allowed. Paths are validated before construction.
+-- 3. The default Path type itself affords considerable safety regarding the
 -- distinction of rooted or non-rooted paths, it also allows distinguishing
 -- directory and file paths.
--- 3. It is designed to provide flexible typing to provide compile time safety
+-- 4. It is designed to provide flexible typing to provide compile time safety
 -- for rooted/non-rooted paths and file/dir paths. The Path type is just part
 -- of that typed path ecosystem. Though the default Path type itself should be
 -- enough for most cases.
--- 4. It leverages the streamly array module for most of the heavy lifting,
+-- 5. It leverages the streamly array module for most of the heavy lifting,
 -- it is a thin wrapper on top of that, improving maintainability as well as
 -- providing better performance. We can have pinned and unpinned paths, also
 -- provide lower level operations for certain cases to interact more
 -- efficiently with low level code.
+-- 6. share name is part of the root when we split the root, this allows us to
+-- treat the server and share name always in cases insensitive manner and the
+-- remaining path can be normalized as case sensitive or insensitive.
 --
 -- It builds on arrays, has a richer API, consistent API, streaming ops where
 -- it makes sense, performance is primary goal.